iox-#1614 Move UninitializedArray to internal, add static_assert
