+++ .clang-tidy-diff-scans.txt
@@ -37,7 +37,6 @@
 ./iceoryx_hoofs/test/moduletests/test_posix*
 ./iceoryx_hoofs/include/iceoryx_hoofs/design_pattern/builder.hpp
 
-./iceoryx_hoofs/include/iceoryx_hoofs/containers/*
 ./iceoryx_hoofs/include/iceoryx_hoofs/internal/containers/*
 ./iceoryx_hoofs/test/moduletests/test_containers_*
