+++ iceoryx_hoofs/include/iceoryx_hoofs/cxx/stack.hpp
@@ -16,9 +16,9 @@
 #ifndef IOX_HOOFS_CXX_STACK_HPP
 #define IOX_HOOFS_CXX_STACK_HPP
 
-#include "iceoryx_hoofs/containers/uninitialized_array.hpp"
 #include "iceoryx_hoofs/cxx/algorithm.hpp"
 #include "iceoryx_hoofs/cxx/optional.hpp"
+#include "iceoryx_hoofs/internal/containers/uninitialized_array.hpp"
 
 #include <cstdint>
